summary
changelog-----
touching the lobster no longer plays the slurp sound effect every frame and instead plays the ow sound and ends the game
lobster now moves
lobster now turns
lobster has a random turn speed every frame it is touching the edge for. this is just a crude way to randomize how the lobster bounces off a wall so that it isn't a predictable pattern.